fix bugs in test_cartepizza class
